Move metadata retrieval logic to MetadataController.
